Move function signatures to header file and separate into two source files.
