+++ Makefile
@@ -6,7 +6,7 @@
 #    By: msousa <mlrcbsousa@gmail.com>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2021/10/29 15:00:20 by msousa            #+#    #+#              #
-#    Updated: 2022/02/23 22:14:53 by msousa           ###   ########.fr        #
+#    Updated: 2022/02/24 22:56:21 by msousa           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -21,6 +21,7 @@ SRC			= minishell.c \
 					src/astree.c \
 					src/tokens.c \
 					src/parser.c \
+					src/parser_1.c \
 					src/utils/binary_paths.c
 
 NAME		= minishell
